add route delete

+++ backend/routes/helper/helper.py
@@ -5,7 +5,7 @@ from pydantic import ValidationError
 from sqlalchemy import inspect
 from sqlalchemy.exc import IntegrityError
 from ORM import Model_description
-from ORM.Model import User, Instance, Phase, InstanceField
+from ORM.Model import User, Instance, Phase, InstanceField, Form, Group, Role, Position
 from ORM.session import Session
 from exceptions import ORMExceptions as ORMExc, InstanceException as InsExc
 from pydantic_models import Schema
@@ -17,7 +17,7 @@ def check_rsc_exist(rsc: str, all_rscs: list):
         raise ORMExc.ResourceNotExists(rsc, all_rscs)
 
 
-def check_auth(schema, current_user: User, rsc_ins: Optional[Any] = None):
+def check_auth(schema, current_user: User, rsc_ins: Optional[Any] = None, action: Optional[str] = None ):
     if hasattr(schema.Config, 'require_admin') and schema.Config.require_admin:
         if not current_user.is_admin:
             raise ORMExc.RequireAdmin
@@ -204,6 +204,15 @@ def upd_rsc_ins(val_dat: dict, rsc_ins: Any, session: Session, current_user: Use
 def del_rsc_ins(rsc_ins: Any):
     session = inspect(rsc_ins).session
     # TODO: delete user, form
+    rsc_ins_type = type(rsc_ins)
+    if rsc_ins_type == User \
+            or (rsc_ins_type == Form and rsc_ins.public) \
+            or (rsc_ins_type == Instance and rsc_ins.current_state != "initialized") \
+            or (rsc_ins_type == Group and rsc_ins.joiners is not None) \
+            or (rsc_ins_type == Role and rsc_ins.holders is not None) \
+            or (rsc_ins_type == Position and rsc_ins.holders is not None):
+        raise ORMExc.IndelibleResourceInstance
+
     try:
         session.delete(rsc_ins)
         session.flush()
